add type alias for native go1.9 monotime
