+++ clock_clock_gettime.go
@@ -1,4 +1,5 @@
 // +build netbsd freebsd solaris
+// +build !go1.9
 
 package monotime
 
@@ -8,7 +9,7 @@ import (
 	"unsafe"
 )
 
-func monotime() uint64 {
+func monotime() Monotime {
 	var ts syscall.Timespec
 
 	syscall.Syscall(syscall.SYS_CLOCK_GETTIME, clock_monotonic, uintptr(unsafe.Pointer(&ts)), 0)
@@ -16,6 +17,6 @@ func monotime() uint64 {
 	return uint64(ts.Nano())
 }
 
-func duration(start uint64, end uint64) time.Duration {
+func duration(start Monotime, end Monotime) time.Duration {
 	return time.Duration(end - start)
 }
