+++ clock_mach_absolute_time.go
@@ -1,4 +1,5 @@
 // +build darwin
+// +build !go1.9
 
 package monotime
 
@@ -15,11 +16,11 @@ func init() {
 	C.mach_timebase_info(&tbinfo)
 }
 
-func monotime() uint64 {
-	return uint64(C.mach_absolute_time())
+func monotime() Monotime {
+	return Monotime(C.mach_absolute_time())
 }
 
-func duration(start uint64, end uint64) time.Duration {
+func duration(start Monotime, end Monotime) time.Duration {
 	elapsed := end - start
 	elapsedNano := elapsed * uint64(tbinfo.numer) / uint64(tbinfo.denom)
 	return time.Duration(elapsedNano)
